requirements.txt
anyio==4.9.0
asgiref==3.8.1
beautifulsoup4==4.13.4
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.7.1
Django==5.2
django-debug-toolbar==5.1.0
django-extensions==4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-nested-routers==0.94.1
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
mysqlclient==2.2.7
oauthlib==3.2.2
pillow==11.2.1
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.1.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
